+++ src/app.js
@@ -21,7 +21,7 @@
   // Middlewares
   app.use(require('morgan')('dev'))
   app.use(require('body-parser').json())
-  app.use(require('./auth')(app, process.env.LOGIN_FILE, process.env.LOGIN, process.env.ACCESS_KEY))
+  app.use(require('./auth')(app, process.env.COOKIE_FILE, process.env.LOGIN, process.env.ACCESS_KEY))
   app.use(require('compression')())
   app.use(require('helmet')({ contentSecurityPolicy: { directives: {
     defaultSrc: [ "'self'" ]